+++ tests/test_suite/test_version.py
@@ -16,4 +16,4 @@ from pyaurorax import __version__
 
 
 def test_version():
-    assert __version__ == "1.23.0"
+    assert __version__ == "1.24.0"